Almost

+++ static/google_maps.js
@@ -1,6 +1,6 @@
 let map;
 
-const bluemarker = "https://maps.google.com/mapfiles/kml/paddle/blu-circle.png"
+const blueMarker = "https://maps.google.com/mapfiles/kml/paddle/blu-circle.png";
 let markers = [];
 
 function initMap() {
@@ -34,13 +34,9 @@ function initMap() {
     let car292indexes = [];
     let car246indexes = [];
     let car34index;
-    let car34LastIndex;
     let car386index;
-    let car386LastIndex;
     let car292index;
-    let car292LastIndex;
     let car246index;
-    let car246LastIndex;
 
     for (let i = 0; i < mydataJson.length; i++) {
         mydataJson[i].date_time = mydataJson[i].date_time.split(" ").slice(-1);
@@ -77,11 +73,6 @@ function initMap() {
         }
     }
 
-    // car34LastIndex = car34indexes[-1];
-    // car386LastIndex = car386indexes[-1];
-    // car292LastIndex = car292indexes[-1];
-    // car246LastIndex = car292indexes[-1];
-
     if (user_id == 2) {
         last30Minutes(user_id, car34index, mydataJson, 0);
         last30Minutes(user_id, car386index, mydataJson, 1);
@@ -93,26 +84,48 @@ function initMap() {
     document.getElementById("submit").onclick = function () {
         let startHour = document.getElementById("startHour").value;
         let endHour = document.getElementById("endHour").value;
-        console.log(startHour)
-        console.log(endHour)
-        deleteMarkers();
+        let carSelect = document.querySelector('#car-select');
+        let car_id = carSelect.value;
 
-        if (user_id == 2) {
-            showRoute(user_id, car34indexes[0], car34indexes[-1], mydataJson, 0, startHour, endHour);
-            showRoute(user_id, car386indexes[0], car386indexes[-1], mydataJson, 1, startHour, endHour)
-        } else if (user_id == 3) {
-            showRoute(user_id, car292indexes[0], car292indexes[-1], mydataJson, 0, startHour, endHour);
-            showRoute(user_id, car246indexes[0], car246indexes[-1], mydataJson, 1, startHour, endHour);
+        if (enforceMinMax(startHour, endHour)) {
+            deleteMarkers();
+
+            if (user_id == 2) {
+
+                if (car_id == 34)
+                    showRoute(user_id, car_id, car34indexes[0], car34indexes[-1], mydataJson, 0, startHour, endHour);
+                else if (car_id == 386)
+                    showRoute(user_id, car_id, car386indexes[0], car386indexes[-1], mydataJson, 1, startHour, endHour)
+            } else if (user_id == 3) {
+
+                if (car_id == 292)
+                    showRoute(user_id, car_id, car292indexes[0], car292indexes[-1], mydataJson, 0, startHour, endHour);
+                else if (car_id == 246)
+                    showRoute(user_id, car_id, car246indexes[0], car246indexes[-1], mydataJson, 1, startHour, endHour);
+                else
+                    console.log(user_id)
+            }
         }
     }
 
 }
+
+
+function enforceMinMax(startHour, endHour) {
+
+    if (-1 < parseInt(startHour) && parseInt(startHour) < 25
+        && -1 < parseInt(endHour) && parseInt(endHour) < 25) {
+        return true;
+    } else {
+        alert("Please enter a number between 0 and 24")
+        return false;
+    }
+}
+
 
 // shows the route for the last 30 minutes
 function last30Minutes(user_id, carIndex, mydataJson, marker_id) {
-
     for (let i = carIndex; (carIndex - 30) < i; i--) {
-        console.log("GİRDİ")
         let coords = {lat: mydataJson[i].latitude, lng: mydataJson[i].longitude}
         let marker;
         if (marker_id === 0) {
@@ -139,29 +152,31 @@ function last30Minutes(user_id, carIndex, mydataJson, marker_id) {
 }
 
 
-function showRoute(user_id, carFirstIndex, carLastIndex, mydataJson, marker_id, startHour, endHour) {
+function showRoute(user_id, car_id, carFirstIndex, carLastIndex, mydataJson, marker_id, startHour, endHour) {
     for (let i = (carFirstIndex + startHour * 60); i < (carFirstIndex + endHour * 60); i++) {
-        let coords = {lat: mydataJson[i].latitude, lng: mydataJson[i].longitude}
-        let marker;
-        if (marker_id === 0) {
-            marker = addMarker(coords);
-        } else if (marker_id === 1) {
-            marker = addBlueMarker(coords);
-        }
+        if (car_id == mydataJson[i].car_id) {
+            let coords = {lat: mydataJson[i].latitude, lng: mydataJson[i].longitude};
+            let marker;
+            if (marker_id === 0) {
+                marker = addMarker(coords);
+            } else if (marker_id === 1) {
+                marker = addBlueMarker(coords);
+            }
 
-        // console.log(typeof (mydataJson[0].latitude));
-        let infowindow = new google.maps.InfoWindow({
-            content: `lat:${coords.lat}, lng:${coords.lng},
+            // console.log(typeof (mydataJson[0].latitude));
+            let infowindow = new google.maps.InfoWindow({
+                content: `lat:${coords.lat}, lng:${coords.lng},
                       date:${mydataJson[i].date_time},
                       car_id:${mydataJson[i].car_id}`
-        });
-        marker.addListener("click", () => {
-            infowindow.open({
-                anchor: marker,
-                map,
-                shouldFocus: false,
             });
-        });
+            marker.addListener("click", () => {
+                infowindow.open({
+                    anchor: marker,
+                    map,
+                    shouldFocus: false,
+                });
+            });
+        }
     }
 
 }
@@ -180,38 +195,27 @@ function addMarker(coords) {
 function addBlueMarker(coords) {
     let marker = new google.maps.Marker({
         position: coords,
-        icon: bluemarker,
+        icon: blueMarker,
         map: map
     });
     markers.push(marker);
     return marker;
-
 }
-
-
 
 // Sets the map on all markers in the array.
 function setMapOnAll(map) {
-  for (let i = 0; i < markers.length; i++) {
-    markers[i].setMap(map);
-  }
+    for (let i = 0; i < markers.length; i++) {
+        markers[i].setMap(map);
+    }
 }
 
 
 // Deletes all markers in the array by removing references to them.
 function deleteMarkers() {
-  setMapOnAll(null);
-  markers = [];
+    setMapOnAll(null);
+    markers = [];
 }
 
 function padTo2Digits(num) {
     return String(num).padStart(2, '0');
 }
-
-
-// function validate(ele) {
-//     var value = ele.value;
-//     var validation = ele.getAttribute("start-time")
-//     console.log(value + "  " + validation);
-//     return false;
-// }
